call Paul, Olivier 2020-05-27
Following our call, I have made the following changes:

App/Powerpoint:

Renamed the powerpoint "Data Management in ACORN App v1.2.0" to have a versioning consistent with the App.
Made the changes listed in slide 2 in the powerpoint and the App.
Microbiology Tab (4): on both numerators and denominators, replaced pull(isolate_id) with pull(specimen_id)
Microbiology Tab (5) & (8): added a filter to remove Mixed growth and Not cultured: filter(organism != "Mixed growth", organism != "Not cultured”)
AMR (1): Added to all counts %>% fun_deduplication(method = input$deduplication_method)
Replaced a screenshot in the powerpoint slide Microbiology Tab.
In the App (this is not data management per se so changes are not in the powerpoint):

Microbiology Tab, (5): replaced top_n(y, n = 10) %>% with arrange(desc(y)) %>% head(10) %>% as top_n can select more than 10 if equal ranks.
Microbiology Tab, (8): reordered the tabel by organism count to match with (5).
Microbiology Tab, (6) & (7): reordered (6) by number of specimens collected (y) instead of the % of specimen collected (freq). The issue was with the rounding and that 10 or 15 specimens amount to the same 1% of all specimens collected with a dataset with a larger number of specimens (denominator).